Allow user to designate build side of the join.
When both sides of the join fit in memory, best join performance is achieved by building an index from the larger side. However, when larger side of the join doesn't fit in memory, the only option is to build an index from the smaller side. This change updates JoinQuery.spatialJoin API to allow the user to choose which side to use when building an index.

Furthermore, the behavior of JoinQuery.SpatialJoinXXX APIs when useIndex = true has been modified as follows:

- if spatialRdd has an index, that index is used (same as before)
- otherwise, if queryRdd has an index, that index is used (threw an exception before),
- otherwise, an ephemeral index on spatialRdd is built on the fly(threw an exception before)