Fix filtering out null data
There is no way to distinguish between having no user data
(in which case it is set to "null") and having one column
of user data that may take on null values (in which case
the string representation would also be "null").

But in the second case, we want to preserve it and not drop
the null values.