[GH-2166] More pathing fixes for Python package refactor (#2167)
* SedonaUtils fix

* maintain backwards compat from the restructuring in #1930

mostly these are just adding "redirects" to the new location.

the geoarrow I moved out of the util folder. unless we had never released this at the time, this seemed to just be a mistake that failed to maintain backwards compatability.

I skipped some things because I thought maybe theyre not supported APIS:
* everything in core
* sedona.register outside of what is available from the init file

I'm sure I missed some things here but hopefull this makes things a lot better for existing users of sedona.

* PR Comments + remove SedonaMapUtils.py, its not a public interface.

* revert test style changes

* clean up imports

* remove print statements

* fix tests

* Fix raster UDT import order issue

---------

Co-authored-by: Kristin Cowalcijk <bo@wherobots.com>