[CI] pre-commit: add hook `chmod` with `manual` stage
Can you run chmod on Windows?

The name is an abbreviation of change mode, which does not exist in Windows OS

Set this hook to manual since chmod doesn't run on Windows.

Allows Windows users to still run the standard pre-commit workflow.

The manual stage can be run by Mac and Linux users.

We can setup the manual stage to run on GitHub actions so we don't lose coverage and there is another PR for that with other hooks that have manual stages.